+++ src/main/java/dk/dtu/compute/course02324/assignment3/lists/uses/PersonsGUI.java
@@ -1,0 +1,156 @@
+package dk.dtu.compute.course02324.assignment3.lists.uses;
+
+
+import dk.dtu.compute.course02324.assignment3.lists.implementations.GenericComparator;
+import dk.dtu.compute.course02324.assignment3.lists.types.List;
+import javafx.geometry.Insets;
+import javafx.geometry.Pos;
+import javafx.scene.control.Button;
+import javafx.scene.control.Label;
+import javafx.scene.control.ScrollPane;
+import javafx.scene.control.TextField;
+import javafx.scene.layout.GridPane;
+import javafx.scene.layout.HBox;
+import javafx.scene.layout.VBox;
+
+import javax.validation.constraints.NotNull;
+import java.util.Comparator;
+
+/**
+ * A GUI element that is allows the user to interact and
+ * change a list of persons.
+ *
+ * @author Ekkart Kindler, ekki@dtu.dk
+ */
+public class PersonsGUI extends GridPane {
+
+    /**
+     * The list of persons to be maintained in this GUI.
+     */
+    final private List<Person> persons;
+
+    private GridPane personsPane;
+
+    private int weightCount = 1;
+
+    /**
+     * Constructor which sets up the GUI attached a list of persons.
+     *
+     * @param persons the list of persons which is to be maintained in
+     *                this GUI component; it must not be <code>null</code>
+     */
+    public PersonsGUI(@NotNull List<Person> persons) {
+        this.persons = persons;
+
+        this.setVgap(5.0);
+        this.setHgap(5.0);
+
+        // text filed for user entering a name
+        TextField field = new TextField();
+        field.setPrefColumnCount(5);
+        field.setText("name");
+
+        // TODO for all buttons installed below, the actions need to properly
+        //      handle (catch) exceptions, and it would be nice if the GUI
+        //      could also show the exceptions thrown by user actions on
+        //      button pressed (cf. Assignment 2).
+
+        // button for adding a new person to the list (based on
+        // the name in the text field (the weight is just incrementing)
+        // TODO a text field for the weight could be added to this GUI
+        Button addButton = new Button("Add");
+        addButton.setOnAction(
+                e -> {
+                    Person person = new Person(field.getText(), weightCount++);
+                    persons.add(person);
+                    // makes sure that the GUI is updated accordingly
+                    update();
+                });
+
+        Comparator<Person> comparator = new GenericComparator<>();
+
+        // button for sorting the list (according to the order of Persons,
+        // which implement the interface Comparable, which is converted
+        // to a Comparator by the GenericComparator above)
+        Button sortButton = new Button("Sort");
+        sortButton.setOnAction(
+                e -> {
+                    persons.sort(comparator);
+                    // makes sure that the GUI is updated accordingly
+                    update();
+                });
+
+        // button for clearing the list
+        Button clearButton = new Button("Clear");
+        clearButton.setOnAction(
+                e -> {
+                    persons.clear();
+                    // makes sure that the GUI is updated accordingly
+                    update();
+                });
+
+        // combines the above elements into vertically arranged boxes
+        // which are then added to the left column of the grid pane
+        VBox actionBox = new VBox(field, addButton, sortButton, clearButton);
+        actionBox.setSpacing(5.0);
+        this.add(actionBox, 0, 0);
+
+        // create the elements of the right column of the GUI
+        // (scrollable person list) ...
+        Label labelPersonsList = new Label("Persons:");
+
+        personsPane = new GridPane();
+        personsPane.setPadding(new Insets(5));
+        personsPane.setHgap(5);
+        personsPane.setVgap(5);
+
+        ScrollPane scrollPane = new ScrollPane(personsPane);
+        scrollPane.setMinWidth(300);
+        scrollPane.setMaxWidth(300);
+        scrollPane.setMinHeight(300);
+        scrollPane.setMaxHeight(300);
+        scrollPane.setVbarPolicy(ScrollPane.ScrollBarPolicy.AS_NEEDED);
+        scrollPane.setHbarPolicy(ScrollPane.ScrollBarPolicy.AS_NEEDED);
+
+        // ... and adds these elements to the right-hand columns of
+        // the grid pane
+        VBox personsList = new VBox(labelPersonsList, scrollPane);
+        personsList.setSpacing(5.0);
+        this.add(personsList, 1, 0);
+
+        // updates the values of the different components with the values from
+        // the stack
+        update();
+    }
+
+    /**
+     * Updates the values of the GUI elements with the current values
+     * from the list.
+     */
+    private void update() {
+        personsPane.getChildren().clear();
+        // adds all persons to the list in the personsPane (with
+        // a delete button in front of it)
+        for (int i=0; i < persons.size(); i++) {
+            Person person = persons.get(i);
+            Label personLabel = new Label(i + ": " + person.toString());
+            Button deleteButton = new Button("Delete");
+            deleteButton.setOnAction(
+                    e -> {
+                        persons.remove(person);
+                        update();
+                    }
+            );
+            HBox entry = new HBox(deleteButton, personLabel);
+            entry.setSpacing(5.0);
+            entry.setAlignment(Pos.BASELINE_LEFT);
+            personsPane.add(entry, 0, i);
+        }
+    }
+
+    // TODO this GUI could be extended by some additional widgets for issuing other
+    //      operations of lists. And the possibly thrown exceptions should be caught
+    //      in the event handler (and possibly shown in an additional text area for
+    //      exceptions; see Assignment 2).
+
+}
