+++ src/main/java/module-info.java
@@ -1,0 +1,15 @@
+/**
+ * Module for assignment3 of the course 02324 in spring 2025.
+ *
+ * @author Ekkart Kindler, ekki@dtu.dk
+ */
+module dk.dtu.compute.course02324.assignment3 {
+
+    requires validation.api;
+    requires javafx.controls;
+
+    exports dk.dtu.compute.course02324.assignment3.lists.types;
+    exports dk.dtu.compute.course02324.assignment3.lists.implementations;
+    exports dk.dtu.compute.course02324.assignment3.lists.uses;
+
+}
